--- user-contract-mon/.gitignore
@@ -1,1 +1,0 @@
-.config
