--- user-contract-mon/README.md
@@ -1,1 +1,0 @@
-# user-contract-mon
